@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.2"
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "com.yalantis.ucrop.sample"
-        minSdkVersion 19
-        targetSdkVersion 24
+        minSdkVersion 23
+        targetSdkVersion 30
         versionCode 10
         versionName "1.2.2"
     }
@@ -18,17 +17,20 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 	lintOptions {
         abortOnError false
     }
-	
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:24.2.0'
+    implementation 'androidx.exifinterface:exifinterface:1.3.2'
 
-    compile project (':ucrop')
+    implementation project (':ucrop')
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
 }
