@@ -16,12 +16,11 @@ apply plugin: 'com.android.library'
 apply from: '../mavenpush.gradle'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '24.0.3'
+    compileSdkVersion 30
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 24
+        minSdkVersion 23
+        targetSdkVersion 30
         versionCode 22
         versionName "2.2.0"
 
@@ -42,15 +41,15 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:25.2.0@aar'
-    compile 'com.android.support:design:25.2.0@aar'
-    compile 'com.android.support:support-v4:25.2.0@aar'
-    compile 'com.android.support:cardview-v7:25.2.0@aar'
-    compile 'com.android.support:percent:25.2.0@aar'
-    compile 'uk.co.chrisjenx:calligraphy:2.2.0'
-    compile 'com.squareup.okhttp3:okhttp:3.4.1'
+    implementation 'androidx.exifinterface:exifinterface:1.3.2'
 
-    compile 'com.shawnlin:number-picker:2.4.3'
-    compile 'com.wdullaer:materialdatetimepicker:3.2.1'
-    compile 'com.github.Kraskovskiy:PopupMenuView:dev3-SNAPSHOT'
+    implementation 'io.github.inflationx:calligraphy3:3.1.1'
+    implementation 'io.github.inflationx:viewpump:2.0.3'
+
+    implementation 'com.shawnlin:number-picker:2.4.3'
+    implementation 'com.wdullaer:materialdatetimepicker:3.2.1'
+    implementation 'com.github.Kraskovskiy:PopupMenuView:dev3-SNAPSHOT'
+
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
 }
