@@ -1,11 +1,13 @@
 package com.yalantis.ucrop.sample;
 
 import android.content.DialogInterface;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+
 
 /**
  * Created by Oleksii Shliama (https://github.com/shliama).
@@ -37,13 +39,8 @@ public class BaseActivity extends AppCompatActivity {
     protected void requestPermission(final String permission, String rationale, final int requestCode) {
         if (ActivityCompat.shouldShowRequestPermissionRationale(this, permission)) {
             showAlertDialog(getString(R.string.permission_title_rationale), rationale,
-                    new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            ActivityCompat.requestPermissions(BaseActivity.this,
-                                    new String[]{permission}, requestCode);
-                        }
-                    }, getString(R.string.label_ok), null, getString(R.string.label_cancel));
+                    (dialog, which) -> ActivityCompat.requestPermissions(BaseActivity.this,
+                            new String[]{permission}, requestCode), getString(R.string.label_ok), null, getString(R.string.label_cancel));
         } else {
             ActivityCompat.requestPermissions(this, new String[]{permission}, requestCode);
         }
