@@ -1,7 +1,9 @@
 package com.yalantis.ucrop.callback;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
+
 
 public interface BitmapCropCallback {
 
