@@ -1,8 +1,10 @@
 package com.yalantis.ucrop.callback;
 
 import android.graphics.Bitmap;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.yalantis.ucrop.model.ExifInfo;
 
